+++ NEWS.md
@@ -1,6 +1,7 @@
 # rp5pik 0.0.2
 
 * Minor improvements in `parse_pik` description and dependencies.
+* Added function `aggregate_pik` to summarise meteo data on a daily or 12-hours basis
 * New function `get_temp50` for detecting a 50% rain-snow air temperature treshold based on the work by *Jennings et al. ([2018](https://www.nature.com/articles/s41467-018-03629-7))*